feat(source error): Adds error handling policy `ignore` to sources
refs #96
